feat: upgrade Phoenix Rise SDK — remove deprecated python/, add ts/ + docs/
QA verified 73/77 tests pass. 4 pre-existing failures (backend endpoint issues,
on-chain program error 7002 for isolated TP/SL) — no regressions from upgrade.

Rust SDK updates: new API module, cancel_stop_loss, multi_limit_order,
register_trader examples, env/http_client refactor. Python SDK removed
(no longer maintained upstream). TypeScript SDK and docs added.

+++ phoenix-rise-sdk/rust/math/src/funding.rs
@@ -75,7 +75,7 @@ impl FundingCalculator {
         // using the i128 max to match on-chain bounds exactly.
         let projected_f = acc_i128 as f64 / period_i128 as f64; // quote lots per base lot
         let max_rate = self.max_funding_rate_per_interval.as_inner() as f64;
-        let clamped = projected_f.clamp(-max_rate as f64, max_rate as f64);
+        let clamped = projected_f.clamp(-max_rate, max_rate);
 
         // Convert to quote units per base unit.
         let quote_lots_per_quote_unit = 10f64.powi(self.quote_lot_decimals as i32);
